requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.4.0
bootstrap-admin==0.4.4
celery==5.1.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
conda==4.3.16
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.2
django-admin-interface==0.14.0
django-celery-beat==2.2.1
django-colorfield==0.3.2
django-dotenv==1.4.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.4.0
-e git+https://github.com/molinae/django-taggit.git@1520409680461e23b65dc379339726fe69980389#egg=django_taggit
django-timezone-field==4.1.2
django-upload-path==0.1.1
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
jdcal==1.4.1
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
pipenv==2020.8.13
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycosat==0.6.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
